+++ packages/tests/tests/waku.node.spec.ts
@@ -1,6 +1,6 @@
+import { bootstrap } from "@libp2p/bootstrap";
 import type { PeerId } from "@libp2p/interface-peer-id";
 import { bytesToUtf8, utf8ToBytes } from "@waku/byte-utils";
-import { PeerDiscoveryStaticPeers } from "@waku/core/lib/peer_discovery_static_list";
 import { waitForRemotePeer } from "@waku/core/lib/wait_for_remote_peer";
 import { createLightNode, createPrivacyNode } from "@waku/create";
 import type {
@@ -71,7 +71,7 @@ describe("Waku Dial [node only]", function () {
       waku = await createLightNode({
         staticNoiseKey: NOISE_KEY_1,
         libp2p: {
-          peerDiscovery: [new PeerDiscoveryStaticPeers([multiAddrWithId])],
+          peerDiscovery: [bootstrap({ list: [multiAddrWithId.toString()] })],
         },
       });
       await waku.start();
@@ -94,12 +94,12 @@ describe("Waku Dial [node only]", function () {
       nwaku = new Nwaku(makeLogFileName(this));
       await nwaku.start();
 
+      const nwakuMa = await nwaku.getMultiaddrWithId();
+
       waku = await createLightNode({
         staticNoiseKey: NOISE_KEY_1,
         libp2p: {
-          peerDiscovery: [
-            new PeerDiscoveryStaticPeers([await nwaku.getMultiaddrWithId()]),
-          ],
+          peerDiscovery: [bootstrap({ list: [nwakuMa.toString()] })],
         },
       });
       await waku.start();